editmode bugfix

+++ functions/backend/products/update_do.php
@@ -1,5 +1,5 @@
 <?php
-
+session_start();
 $ean = $_POST["ean"];
 $name = $_POST["name"];
 $beschreibung = $_POST["beschreibung"];
@@ -7,7 +7,11 @@ $preis = $_POST["preis"];
 $rating = $_POST["rating"];
 $genre = $_POST["genre"];
 $oldbild = $_POST["oldbild"];
-$produktbild = $_FILES['bild']['name']; //Name der Bilddatei mit Endung
+if (strlen ($_FILES['bild']['name']) > 0) {
+	$produktbild = $_FILES['bild']['name'];
+} else {
+	$produktbild = $oldbild;
+}
 $filename = pathinfo($_FILES['bild']['name'], PATHINFO_FILENAME);
 $extension = strtolower(pathinfo($_FILES['bild']['name'], PATHINFO_EXTENSION));
 
@@ -21,20 +25,26 @@ if (!empty($ean) && !empty($name) && !empty($beschreibung) && !empty($preis) && 
             "UPDATE sortiment SET name= :name, beschreibung= :beschreibung, genre= :genre, preis= :preis, rating= :rating, bild= :bild WHERE ean= :ean");
         $query->execute(array("name" => $name, "beschreibung" => $beschreibung, "genre" => $genre, "preis" => $preis, "rating" => $rating, "bild" => $produktbild, "ean" => $ean));
         $db = null;
-        header('Location: ../../../admin.php');
+		header("Location: ../../../admin.php");
     } catch (PDOException $e) {
-        echo "Error!";
-        die();
+		$_SESSION["error"] = "true";
+    	die();
     }
 } else {
-    echo "Error: Bitte alle Felder ausfüllen!";
+	header("Location: ../../../admin.php");
+	$_SESSION["error"] = "true";
+	die();
 }
 
 if(in_array($extension, $allowed_extensions)) {
     move_uploaded_file($_FILES['bild']['tmp_name'], '../../../files/uploads/'.$_FILES['bild']['name']);
+	if ($oldbild !== $produktbild) {
 	unlink ('../../../files/uploads/'.$oldbild);
-    header("Location: ../../../admin.php");
+	header("Location: ../../../admin.php");
+	}
 }
 else {
+	$_SESSION["error"] = "true";
+	header("Location: ../../../admin.php");
     die("Ungültige Dateiendung. Nur png, jpg, jpeg und gif-Dateien sind erlaubt");
 }
